+++ net/processor.go
@@ -2,13 +2,13 @@ package net
 
 type Processor interface {
 	//发送处理
-	Route(id uint16, msg interface{}, userData *UserData) error
+	Route(id uint32, msg interface{}, userData *UserData) error
 	//解包数据
-	Unmarshal(data []byte) (uint16, interface{}, error)
+	Unmarshal(data []byte) (uint32, interface{}, error)
 	//打包数据
-	Marshal(msg interface{}) (uint16, [][]byte, error)
+	Marshal(msg interface{}) (uint32, [][]byte, error)
 
-	Register(msg interface{}, id uint16) uint16
+	Register(msg interface{}, id uint32) uint32
 
-	GetMsg(id uint16) interface{}
+	GetMsg(id uint32) interface{}
 }
